Allows for PICs to have two config words
This is currently only set for the PIC16F886
but should be set for all double config PICs.

The trigger for this is to add 0x10000 to the PIC's
protocol_id in devices.h. This is stripped off in
main.c after setting the correct PIC fuse structure.

For compatibility, the original config word (0x2007) is
still conf_word in the input/output text file.

The second config word (0x2008) is tagged conf_word1

Both read and write will respect the extra configuration word.

Note the Makefile doesn't pick up changes to devices.h so if
you change more PICs to handle dual config words, you need to
do a make clean before running make to be sure the changes are picked
up.