Add support for disabling/enabling protection
The original software disables write protection before write and restores it
after write if the chip supports it. It provides a checkboxes for disabling the
functionality.

This changeset adds equivalent functionality with equivalent defaults. The bits
in opts4 were guessed by clicking around in the UI and seeing whether the
checkboxes grey out. The opcodes 0x44 and 0x45 and the transaction behavior
were identified by sniffing the traffic via NetPacket::USBMon.