Revert "Using $PREFIX instead of the $DESTDIR confusion."
This is mostly wrong. DESTDIR is conventionally used to override the
install root when doing an out-of-tree install (such as on a package
build), while PREFIX denotes the actual install tree (e.g. /usr vs.
/usr/local).

Also the commit changed the udev rules dir to /etc, which is meant for
*user* configuration, not configuration shipped with a package.

A commit that adds PREFIX properly will follow.

This reverts commit 455dd5b5b135eb99f2717d68ad8b1b8cbab97c32.