Implement idcheck_only command (-d option).
Allows to just read the chip id and do nothing else (like reading the
whole chip contents).

This is for use in scripts where you have for example several similar
chips with different IDs and want to check if the currently inserted
chip is one in your list of allowed ones. Also useful to detect if
a chip is inserted at all when using the -y option.