Add build-time Makefile configuration option to add MacOS-specific stuff.
At least with MacOS 10.13, it's not necessary to explicitly link the two
frameworks: Foundation and IOKit.  When libusb-1.0.0.dylib as installed
through MacPorts and Brew, these two frameworks will automatically be
included.  It's not clear which earlier releases require the frameworks
to be explicitly included, so there's now an option in the Makefile to
do this.